refactor(DTFS2-6626): move portal-api role-validator to roles folder
